backend/Ozon-Bank-Manager-Trainer/backend/pyproject.toml
[tool.poetry]
name = "operator-voice-trainer-backend"
version = "0.1.0"
description = "Backend service for Operator Voice Trainer"
authors = ["sssay10 <duvakin.k@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.122.0"
uvicorn = {extras = ["standard"], version = "^0.38.0"}
pydantic = "^2.0.0"
httpx = ">=0.24.0"
langchain = "^0.1.0"
langchain-openai = "^0.1.0"
langchain-core = "^0.1.0"
# Database dependencies
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
ollama = "^0.6.1"
greenlet = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

